Implement test stubs and moderator/problem class tests
This commit primarily stubs out the test files for all current existing
core source packages.

It implements a very simple set of tests for the dextoxify moderator
impelementation.

It also provides an initial stab at providing unit tests for the Problem
class. This is accomplished by providing a dummy implementation of a
Problem in the test file, then confirming that the validation works.